Don't truncate pointer, coerce it to intptr_t instead.

+++ ficl/utility.c
@@ -9,7 +9,7 @@
 void *ficlAlignPointer(void *ptr)
 {
 #if FICL_PLATFORM_ALIGNMENT > 1
-	int p = (int)ptr;
+	intptr_t p = (intptr_t)ptr;
 	if (p & (FICL_PLATFORM_ALIGNMENT - 1))
 		ptr = (void *)((p & ~(FICL_PLATFORM_ALIGNMENT - 1)) + FICL_PLATFORM_ALIGNMENT);
 #endif
